requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
discord==2.3.2
discord.py==2.5.2
fonttools==4.56.0
fortune-python==1.1.1
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.25.1
urllib3==2.3.0
yarl==1.18.3
